function that executes a function given as a parameter on each element of an Array

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -9,7 +9,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned long int Q = 0;
+	size_t Q = 0;
 
 	while (size > Q)
 	{
